Adding Github icons
