+++ assets/i18n/en.json
@@ -1,6 +1,4 @@
 {
-  "about_us_header_label": "About",
-  "about_us_link_label": "About",
   "admin_drawer_link_label": "Admin",
   "admin_header_label": "Admin",
   "alerts_header_label": "Alerts",
@@ -28,6 +26,7 @@
   "features_header_label": "Features",
   "features_link_label": "Features",
   "general_error_snackbar_label": "Oops, something went wrong",
+  "github_link_label": "Github Repository",
   "go_back_link_label": "Go back",
   "go_back_home_link_label": "Go back to home",
   "go_back_profile_link_label": "Go back to profile",
