+++ assets/i18n/nl.json
@@ -1,6 +1,4 @@
 {
-  "about_us_header_label": "Over ons",
-  "about_us_link_label": "Over ons",
   "auth_link_label": "Inloggen",
   "admin_drawer_link_label": "Admin",
   "admin_header_label": "Admin",
@@ -28,6 +26,7 @@
   "features_header_label": "Features",
   "features_link_label": "Features",
   "general_error_snackbar_label": "Oops, er is iets verkeerd gegaan",
+  "github_link_label": "Github Repository",
   "go_back_link_label": "Ga terug",
   "go_back_home_link_label": "Ga terug naar home",
   "go_back_profile_link_label": "Ga terug naar profiel",
