+++ lib/constants/icon_buttons/theme_header_icon_button.dart
@@ -7,24 +7,15 @@ import 'package:responsive_framework/responsive_framework.dart';
 
 import '../../services/theme_service.dart';
 
-class ThemeHeaderIcon extends StatelessWidget {
-  const ThemeHeaderIcon({super.key});
+class ThemeHeaderIconButton extends StatelessWidget {
+  const ThemeHeaderIconButton({super.key});
 
   @override
   Widget build(BuildContext context) {
     return pv.Consumer<ThemeService>(
         builder: (context, theme, child) => theme.darkTheme == true
             ? Padding(
-                padding: EdgeInsets.fromLTRB(
-                    20.0,
-                    10.0,
-                    ResponsiveValue(context,
-                            defaultValue: 50.0,
-                            valueWhen: const [
-                              Condition.smallerThan(name: TABLET, value: 10.0)
-                            ]).value ??
-                        50.0,
-                    0.0),
+                padding: const EdgeInsets.fromLTRB(10.0, 10.0, 10.0, 0.0),
                 child: IconButton(
                     icon: Icon(
                       (defaultTargetPlatform == TargetPlatform.iOS ||
