+++ lib/screens/public/auth_screen.dart
@@ -10,12 +10,12 @@ import 'package:supabase_flutter/supabase_flutter.dart';
 
 import '../../constants/drop_downs/language_header_dropdown_widget.dart';
 import '../../constants/headers/public_menu_header.dart';
+import '../../constants/icon_buttons/github_icon_button.dart';
 import '../../constants/icons/email_icon.dart';
-import '../../constants/icons/public_menu_icon.dart';
-import '../../constants/icons/theme_header_icon.dart';
+import '../../constants/icon_buttons/public_menu_icon_button.dart';
+import '../../constants/icon_buttons/theme_header_icon_button.dart';
 
-import '../../constants/links/about_drawer_link.dart';
-import '../../constants/links/about_header_link.dart';
+import '../../constants/links/github_drawer_link.dart';
 import '../../constants/links/faq_drawer_link.dart';
 import '../../constants/links/faq_header_link.dart';
 import '../../constants/links/features_drawer_link.dart';
@@ -197,7 +197,7 @@ class _AuthScreenState extends State<AuthScreen> {
             FeaturesDrawerLink(highlight: false),
             PricingDrawerLink(highlight: false),
             FaqDrawerLink(highlight: false),
-            AboutDrawerLink(highlight: false)
+            GithubDrawerLink()
           ],
         ),
       );
@@ -219,7 +219,7 @@ class _AuthScreenState extends State<AuthScreen> {
           child: Row(
             mainAxisAlignment: MainAxisAlignment.start,
             crossAxisAlignment: CrossAxisAlignment.center,
-            children: const <Widget>[PublicMenuIcon(), LogoHeaderLink()],
+            children: const <Widget>[PublicMenuIconButton(), LogoHeaderLink()],
           ),
         ),
         titleSpacing: 0,
@@ -228,10 +228,10 @@ class _AuthScreenState extends State<AuthScreen> {
         actions: const [
           FeaturesHeaderLink(highlight: false),
           PricingHeaderLink(highlight: false),
-          AboutUsHeaderLink(highlight: false),
           FaqHeaderLink(highlight: false),
           LanguageHeaderDropdown(),
-          ThemeHeaderIcon(),
+          ThemeHeaderIconButton(),
+          GithubIconButton(),
         ],
       ),
       body: SingleChildScrollView(
