+++ lib/services/util_service.dart
@@ -1,0 +1,18 @@
+import 'package:supabase_flutter/supabase_flutter.dart';
+import 'package:url_launcher/url_launcher.dart';
+
+final supabase = Supabase.instance.client;
+
+class UtilService {
+  final Uri url = Uri.parse('https://github.com/simplism-io');
+
+  Future<void> launchSimplismGithub() async {
+    if (!await launchUrl(
+      url,
+      mode: LaunchMode.inAppWebView,
+      webViewConfiguration: const WebViewConfiguration(enableDomStorage: false),
+    )) {
+      throw Exception('Could not launch $url');
+    }
+  }
+}
