+++ pubspec.yaml
@@ -29,6 +29,7 @@ dependencies:
   file_picker: ^5.2.5
   quill_html_editor: ^2.0.5
   flutter_widget_from_html: ^0.9.1
+  url_launcher: ^6.1.9
 
 dev_dependencies:
   flutter_lints: ^2.0.1
